fix: renamed status_codes.go to statuscodes.go
